backend/requirements.txt
fastapi
uvicorn
sqlalchemy==2.0.31
alembic
pydantic>=2.0.0
pydantic[email]>=2.0.0
pyjwt
python-jose
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv
psycopg2-binary
